Fix int overflow on 8-bit boards analog map example

+++ examples/Analog_map_example/Analog_map_example.ino
@@ -32,7 +32,7 @@ void setup(){
   USBMidiController.setDelay(15);  // wait 15 ms after each message not to flood the connection
   delay(1000); // Wait a second...
   fader.average(analogAverage); // Use the average of 8 samples to get smooth transitions and prevent noise
-  fader.map([](int value){return value*value/1023;}); // Map to a quadratic curve
+  fader.map([](int value){return (int)((unsigned long)value*value/1023);}); // Map to a quadratic curve
 }
 
 //________________________________________________________________________________________________________________________________
